requirements.txt
langchain
langchain-community
langchain-openai
langchain-core

python-dotenv

faiss-cpu

langchain-pinecone
pinecone
pypdf
streamlit